Added example test classes.
